Revert "Trabajando en el Modelo"
This reverts commit 69d7ec87e1417fc221929132f44f879f764e14f4.
