+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "Control-Pcduino",
+  "name": "chat-example",
   "version": "0.0.0",
   "description": "A chat example to showcase how to use `socket.io` with a static `express` server with `async` for control flow.",
   "main": "server.js",
@@ -7,6 +7,7 @@
   "author": "Mostafa Eweda <mo.eweda@gmail.com>",
   "dependencies": {
     "express": "^4.10.2",
+    "pcduino": "0.0.6",
     "socket.io": "^1.0.6"
   }
 }
